requirements.txt
# AgentOS Requirements
# ===================
# Core dependencies for AgentFIN

# Configuration
python-dotenv==1.2.1
requests>=2.31.0

# Orchestration
langgraph>=0.0.10


# Data validation
pydantic==2.12.5
pydantic-ai==1.51.0

# YAML parsing (for SKILL.md)
PyYAML==6.0.3

# Database
# (sqlite3 is built-in to Python)

# Vector Database (for cold memory)
lancedb>=0.5.0


# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# Optional: Development tools
# black==24.10.0
# ruff==0.8.4
